consolidate Tabs source into a single file
